# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DigiProducts API — single-tenant AI digital product creation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /learn/lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [learn]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Lesson detail with progress and next lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
  /learn/lessons/{lessonId}/progress:
    post:
      operationId: postLessonProgress
      tags: [learn]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonProgressInput"
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"
  /learn/lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [learn]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Lesson completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnStatus"
...
components:
  schemas:
...
    Lesson:
      type: object
      required: [id, moduleId, title, orderIndex, durationSeconds, isRequiredForOnboarding, isPublished, allowManualComplete, resources]
      properties:
        id: { type: string }
        moduleId: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        bodyMd: { type: ["string", "null"] }
        videoProvider: { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        durationSeconds: { type: integer }
        orderIndex: { type: integer }
        isRequiredForOnboarding: { type: boolean }
        allowManualComplete: { type: boolean }
        isPublished: { type: boolean }
        locked: { type: boolean }
        resources:
          type: array
          items:
            $ref: "#/components/schemas/LessonResource"
        progress:
          $ref: "#/components/schemas/LessonProgress"
    LessonResource:
      type: object
      required: [id, label]
      properties:
        id: { type: string }
        label: { type: string }
        externalUrl: { type: ["string", "null"] }
    LessonDetail:
      type: object
      required: [lesson]
      properties:
        lesson:
          $ref: "#/components/schemas/Lesson"
        nextLesson:
          $ref: "#/components/schemas/Lesson"
    LessonProgress:
      type: object
      required: [lessonId, status, watchedSeconds, furthestPosition]
      properties:
        lessonId: { type: string }
        status: { type: string, enum: [not_started, in_progress, completed] }
        watchedSeconds: { type: integer }
        furthestPosition: { type: integer }
        completedAt: { type: ["string", "null"] }
    LessonProgressInput:
      type: object
      required: [watchedSeconds, furthestPosition]
      properties:
        watchedSeconds: { type: integer }
        furthestPosition: { type: integer }
    LearnStatus:
      type: object
      required: [requiredCount, completedCount, percentComplete, onboardingComplete]
      properties:
        requiredCount: { type: integer }
        completedCount: { type: integer }
        percentComplete: { type: integer }
        onboardingComplete: { type: boolean }
        justCompleted: { type: boolean }
        nextLessonId: { type: ["string", "null"] }
...