# lib/api-zod/src/generated/types/lessonResource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DigiProducts API — single-tenant AI digital product creation platform
 * OpenAPI spec version: 0.1.0
 */

export interface LessonResource {
  id: string;
  label: string;
  /** @nullable */
  externalUrl?: string | null;
}